+++ app/controllers/profile/UkimsNumberController.scala
@@ -19,7 +19,7 @@ package controllers.profile
 import connectors.TraderProfileConnector
 import controllers.BaseController
 import controllers.actions._
-import forms.UkimsNumberFormProvider
+import forms.profile.UkimsNumberFormProvider
 import models.Mode
 import navigation.ProfileNavigator
 import pages.profile.{UkimsNumberPage, UkimsNumberUpdatePage}
