+++ app/forms/profile/NirmsNumberFormProvider.scala
@@ -14,17 +14,22 @@
  * limitations under the License.
  */
 
-package forms
+package forms.profile
+
+import forms.mappings.Mappings
+import forms.mappings.helpers.FormatAnswers.{addHyphensToNirms, toUppercaseAndRemoveSpacesAndHyphens}
+import models.StringFieldRegex
+import play.api.data.Form
 
 import javax.inject.Inject
 
-import forms.mappings.Mappings
-import play.api.data.Form
+class NirmsNumberFormProvider @Inject() extends Mappings {
 
-class RemoveNirmsFormProvider @Inject() extends Mappings {
-
-  def apply(): Form[Boolean] =
+  def apply(): Form[String] =
     Form(
-      "value" -> boolean("removeNirms.error.required")
+      "value" -> text("nirmsNumber.error.required")
+        .transform(toUppercaseAndRemoveSpacesAndHyphens, identity[String])
+        .verifying(regexp(StringFieldRegex.nirmsRegex, "nirmsNumber.error.invalidFormat"))
+        .transform(addHyphensToNirms, identity[String])
     )
 }
