+++ test/forms/profile/HasNirmsFormProviderSpec.scala
@@ -14,17 +14,32 @@
  * limitations under the License.
  */
 
-package forms
+package forms.profile
 
-import forms.mappings.Mappings
-import play.api.data.Form
+import forms.behaviours.BooleanFieldBehaviours
+import play.api.data.FormError
 
-import javax.inject.Inject
+class HasNirmsFormProviderSpec extends BooleanFieldBehaviours {
 
-class UseExistingUkimsNumberFormProvider @Inject() extends Mappings {
+  val requiredKey = "hasNirms.error.required"
+  val invalidKey  = "error.boolean"
 
-  def apply(): Form[Boolean] =
-    Form(
-      "value" -> boolean("existingUkimsNumber.error.required")
+  val form = new HasNirmsFormProvider()()
+
+  ".value" - {
+
+    val fieldName = "value"
+
+    behave like booleanField(
+      form,
+      fieldName,
+      invalidError = FormError(fieldName, invalidKey)
     )
+
+    behave like mandatoryField(
+      form,
+      fieldName,
+      requiredError = FormError(fieldName, requiredKey)
+    )
+  }
 }
