+++ test/forms/profile/RemoveNiphlFormProviderSpec.scala
@@ -14,20 +14,32 @@
  * limitations under the License.
  */
 
-package forms
+package forms.profile
 
-import javax.inject.Inject
-import forms.mappings.Mappings
-import forms.mappings.helpers.FormatAnswers.toUppercaseAndRemoveSpacesAndHyphens
-import models.StringFieldRegex
-import play.api.data.Form
+import forms.behaviours.BooleanFieldBehaviours
+import play.api.data.FormError
 
-class UkimsNumberFormProvider @Inject() extends Mappings {
+class RemoveNiphlFormProviderSpec extends BooleanFieldBehaviours {
 
-  def apply(): Form[String] =
-    Form(
-      "value" -> text("ukimsNumber.error.required")
-        .transform(toUppercaseAndRemoveSpacesAndHyphens, identity[String])
-        .verifying(regexp(StringFieldRegex.ukimsNumberRegex, "ukimsNumber.error.invalidFormat"))
+  val requiredKey = "removeNiphl.error.required"
+  val invalidKey  = "error.boolean"
+
+  val form = new RemoveNiphlFormProvider()()
+
+  ".value" - {
+
+    val fieldName = "value"
+
+    behave like booleanField(
+      form,
+      fieldName,
+      invalidError = FormError(fieldName, invalidKey)
     )
+
+    behave like mandatoryField(
+      form,
+      fieldName,
+      requiredError = FormError(fieldName, requiredKey)
+    )
+  }
 }
