+++ test/forms/profile/UseExistingUkimsNumberFormProviderSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package forms
+package forms.profile
 
 import forms.behaviours.BooleanFieldBehaviours
 import play.api.data.FormError
